requirements.txt
streamlit
pillow
pillow-heif==0.16.0
